+++ configuracion.py
@@ -1,0 +1,6 @@
+CANVAS_ANCHURA = 800
+CANVAS_ALTURA = 500
+TITULO = "Cazaburbujas"
+FONDO = 'darkblue'
+CENTRO_X = CANVAS_ANCHURA / 2
+CENTRO_Y = CANVAS_ALTURA / 2
